# contracts/openapi/identity-api-v1.yaml
openapi: 3.1.0
info:
  title: MiniPay Identity API
  version: 1.3.0
paths:
...
  /api/v1/users/me:
    get:
      summary: Get the authenticated consumer profile
      security: [{minipayOAuth: [identity.profile.read]}]
      x-minipay-audience: consumer-api
      responses:
        '200': {description: Profile, content: {application/json: {schema: {$ref: '#/components/schemas/ProfileView'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
    patch:
      summary: Update nickname and optionally consume an uploaded avatar
      security: [{minipayOAuth: [identity.profile.write]}]
      x-minipay-audience: consumer-api
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/UpdateProfileRequest'}}
      responses:
        '200': {description: Updated profile, content: {application/json: {schema: {$ref: '#/components/schemas/ProfileView'}}}}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
...
  /api/v1/auth/consumer/code/send:
    post:
      summary: Request a consumer login verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerCodeSendRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumerCodeChallenge'
        '429': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /api/v1/auth/consumer/code/verify:
    post:
      summary: Verify a consumer login code and issue a PKCE-bound authorization code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerCodeVerifyRequest'
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumerAuthorizationCode'
        '400': {$ref: '#/components/responses/Problem'}
        '423': {$ref: '#/components/responses/Problem'}
        '429': {$ref: '#/components/responses/Problem'}
  /api/v1/payment-authorizations:
    post:
      summary: Issue one-time payment authorization
      security: [{minipayOAuth: [identity.payment-authorization.write]}]
      x-minipay-audience: consumer-api
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentAuthorizationRequest'
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/PaymentAuthorization'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/users/me/onboarding:
    put:
      summary: Complete first-login consumer initialization
      description: Sets the required nickname and optional user-bound avatar. Payment password and real-name verification are independent capabilities.
      security: [{minipayOAuth: [identity.profile.write]}]
      x-minipay-audience: consumer-api
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConsumerOnboardingRequest'}
      responses:
        '200': {description: Existing completed onboarding, content: {application/json: {schema: {$ref: '#/components/schemas/ConsumerOnboardingResponse'}}}}
        '201': {description: Completed, content: {application/json: {schema: {$ref: '#/components/schemas/ConsumerOnboardingResponse'}}}}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
...
  /internal/v1/payment-authorizations/verify-and-consume:
    post:
      summary: Verify a bound one-time authorization for a trusted Payment service
      security: [{minipayOAuth: [identity.payment-authorization.verify]}]
      x-minipay-audience: identity-internal
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VerifyPaymentAuthorizationRequest'}
      responses:
        '200':
          description: Verified and consumed idempotently for the same subject
          content:
            application/json:
              schema:
                type: object
                required: [authorizationId, valid]
                properties:
                  authorizationId: {type: string, format: uuid}
                  valid: {type: boolean, const: true}
        '422': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    delegatedBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentClientBasic:
      type: http
      scheme: basic
    minipayOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/oauth2/authorize
          tokenUrl: http://localhost:8081/oauth2/token
          scopes:
            identity.profile.read: Read the authenticated consumer profile
            identity.payment-authorization.write: Issue a one-time payment authorization
            identity.payment-authorization.verify: Verify a payment authorization
            identity.consumer-payment-profile.read: Read a collection recipient's safe display profile
            identity.profile.write: Initialize the authenticated consumer profile
            identity.account.manage: Resolve and manage accounts for trusted internal services
  parameters:
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
  schemas:
...
    ConsumerCodeSendRequest:
      type: object
      additionalProperties: false
      required: [mobile, purpose]
      properties:
        mobile: {type: string, pattern: '^1[3-9]\d{9}$'}
        purpose: {type: string, const: LOGIN}
    ConsumerCodeChallenge:
      type: object
      additionalProperties: false
      required: [challengeId, maskedMobile, expiresAt, resendAfterSeconds]
      properties:
        challengeId: {type: string, format: uuid}
        maskedMobile: {type: string}
        expiresAt: {type: string, format: date-time}
        resendAfterSeconds: {type: integer, minimum: 1}
    ConsumerCodeVerifyRequest:
      type: object
      additionalProperties: false
      required: [challengeId, code, clientId, redirectUri, codeChallenge, codeChallengeMethod, deviceId]
      properties:
        challengeId: {type: string, format: uuid}
        code: {type: string, pattern: '^\d{6}$', writeOnly: true}
        clientId: {type: string, const: minipay-android}
        redirectUri: {type: string}
        codeChallenge: {type: string, minLength: 43, maxLength: 128}
        codeChallengeMethod: {type: string, const: S256}
        deviceId: {type: string, minLength: 1, maxLength: 128}
    ConsumerAuthorizationCode:
      type: object
      additionalProperties: false
      required: [authorizationCode, expiresAt, payPasswordSet, onboardingRequired, realNameStatus, realNameVerified]
      properties:
        authorizationCode: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
        payPasswordSet: {type: boolean}
        onboardingRequired: {type: boolean}
        realNameStatus: {type: string, enum: [UNVERIFIED, PROCESSING, VERIFIED, REJECTED]}
        realNameVerified: {type: boolean}
    PaymentAuthorizationRequest:
      type: object
      additionalProperties: false
      required: [subjectType, subjectId, amountCent, deviceId, payPassword]
      properties:
        subjectType: {type: string, enum: [TRANSFER_INTENT, PAYMENT_ORDER, WITHDRAWAL_ORDER, RECHARGE_ORDER, BANK_CARD_BALANCE_QUERY]}
        subjectId: {type: string, format: uuid}
        amountCent:
          type: integer
          minimum: 0
          maximum: 1000000
          description: Must be 0 only for BANK_CARD_BALANCE_QUERY; all funding subjects require at least 1.
        deviceId: {type: string, minLength: 1, maxLength: 128}
        payPassword: {type: string, pattern: '^\d{6}$', writeOnly: true}
    PaymentAuthorization:
      type: object
      required: [authorizationId, paymentAuthToken, expiresAt]
      properties:
        authorizationId: {type: string, format: uuid}
        paymentAuthToken: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
    VerifyPaymentAuthorizationRequest:
      type: object
      additionalProperties: false
      required: [paymentAuthToken, userId, subjectType, subjectId, amountCent, deviceId]
      properties:
        paymentAuthToken: {type: string, writeOnly: true}
        userId: {type: string, format: uuid}
        subjectType: {type: string, enum: [TRANSFER_INTENT, PAYMENT_ORDER, WITHDRAWAL_ORDER, RECHARGE_ORDER, BANK_CARD_BALANCE_QUERY]}
        subjectId: {type: string, format: uuid}
        amountCent: {type: integer, minimum: 0, maximum: 1000000}
        deviceId: {type: string, minLength: 1, maxLength: 128}
...
    ConsumerOnboardingRequest:
      type: object
      additionalProperties: false
      required: [nickname]
      properties:
        nickname: {type: string, minLength: 2, maxLength: 20}
        avatarUploadId: {type: string, format: uuid}
...
    ConsumerOnboardingResponse:
      type: object
      additionalProperties: false
      required: [userId, nickname, payPasswordSet, onboardingCompleted]
      properties:
        userId: {type: string, format: uuid}
        nickname: {type: string}
        payPasswordSet: {type: boolean}
        onboardingCompleted: {type: boolean}
...
    ProfileView:
      type: object
      additionalProperties: false
      required: [userId, nickname, miniPayNo, version]
      properties:
        userId: {type: string, format: uuid}
        nickname: {type: string, minLength: 2, maxLength: 20}
        miniPayNo: {type: string, maxLength: 32}
        avatarUrl: {type: string, format: uri}
        avatarUrlExpiresAt: {type: string, format: date-time}
        version: {type: integer, format: int64, minimum: 0}
...
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [nickname, version]
      properties:
        nickname: {type: string, minLength: 2, maxLength: 20}
        avatarUploadId: {type: string, format: uuid}
        version: {type: integer, format: int64, minimum: 0}
    Problem: {$ref: './components/problem-details-v1.yaml#/components/schemas/ProblemDetails'}
  responses:
    Problem: {description: Problem Details, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}